chore: :hammer: add checklist to deployment script

+++ deploy.sh
@@ -6,6 +6,23 @@ YELLOW='\033[1;33m'
 RED='\033[0;31m'
 NC='\033[0m' # No Color
 
+
+# Step 0: Checklist
+echo -e "${YELLOW}Step 0: Checklist...${NC}"
+read -p "Did you update your .envs? (y/n)" -n 1 -r
+echo
+if [[ ! $REPLY =~ ^[Yy]$ ]]
+then
+    echo -e "${RED}Update your .envs before deploying.${NC}"
+    exit 1
+fi
+read -p "Did you update your Dockerfiles? (y/n) " -n 1 -r
+echo
+if [[ ! $REPLY =~ ^[Yy]$ ]]
+then
+    echo -e "${RED}Update your Dockerfiles before deploying.${NC}"
+    exit 1
+fi
 # Step 1: Build docker images
 echo -e "${YELLOW}Step 1: Building docker images...${NC}"
 docker rmi $(docker images -q)
